fix: bug with modal and rerender
